# Refs #28459 -- Improved performance of select_related() when model is prefetched from its parent.